Migrate from bs-platform 8.2.0 to the latest rescript 11.1
This change updates the toolchain:

- Replace yarn with npm
- Rename bsconfig/.re to rescript/.res
- Change generic type from `name(a)` to `name<a>`
- Replace `ReasonReact` with `React`
- Replace `Js.Re` with the nicer `RegExp` api.
- Use the new `list{}` syntax in place of `[]`.

The biggest change happens for the code point matcher.
We previously relied on the list{} type, but the warning indicated
that was not efficient in that case.
The code points are now parsed into an array, and we use `get()`
to pattern match without using a `...` spread.

The benchmark test is about 20% faster now:

> $ time node tests/Spec.res.js
> real	0m3.263s

Previously:

> $ time node tests/Spec.bs.js
> real	0m4.192s

Change-Id: I59bf56c856459f8112949f0c93bc3e030ee4d8f5